api/requirements.txt
flask>=3.0
flask-cors>=4.0
pyjwt>=2.8
bcrypt>=4.1
